LayoutTests/http/tests/navigation/ping-cross-origin-from-https.html

<html><head>
<title>Ping</title>
<script>

if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
}

if (location.protocol != "https:")
    location = "https://127.0.0.1:8443/navigation/ping-cross-origin-from-https.html"

var testCalled = false;

function test() {
    if (!testCalled) {
        if (window.testRunner) {
            testRunner.dumpAsText();
            testRunner.overridePreference("WebKitHyperlinkAuditingEnabled", 1);
            testRunner.waitUntilDone();
        }
        testCalled = true;
        return;
    }

    if (window.eventSender) {
        var a = document.getElementById("a");
        eventSender.mouseMoveTo(a.offsetLeft + 2, a.offsetTop + 2);
        eventSender.mouseDown();
        eventSender.mouseUp();
    }
}

</script>
</head>
</script>
</head>
<body onload="test();">
<img src="resources/delete-ping.php" onload="test();" onerror="test();"></img>
<a id="a" href="resources/check-ping.php" ping="http://localhost:8000/navigation/resources/save-Ping.php?test=/navigation/ping-cross-origin-from-https.html">Navigate and send ping</a>
</body></html>
